feat(app): add retry connection modal if there was an error attempting to fetch a passport
if there was an error attempting to fetch a passport, for example if the ceramic connection drops
or timed out, it is unknown whether a user has an existing passport or needs to create a new
passport. this feature adds a modal to describe the situation and allows the user to retry the
connection if desired.

[#196]